moved override test to unit test
